move CronStats file
